Move font picker into ui_input crate from onboarding
