+++ crates/ui_input/Cargo.toml
@@ -14,8 +14,10 @@ path = "src/ui_input.rs"
 [dependencies]
 component.workspace = true
 editor.workspace = true
+fuzzy.workspace = true
 gpui.workspace = true
 menu.workspace = true
+picker.workspace = true
 settings.workspace = true
 theme.workspace = true
 ui.workspace = true
